etls/coaf/requirements.txt
pandas==1.4.3
boto3==1.24.33
requests==2.28.1
psycopg2-binary==2.9.3
lxml==4.9.2
